added weather api
