+++ src/components/Weather.jsx
@@ -1,4 +1,5 @@
 import axios from "axios";
+import { getWeather } from "../utils/lib";
 import { useEffect, useState } from "react";
 
 const API_KEY = import.meta.env.VITE_WEATHER_API_KEY;
@@ -9,26 +10,17 @@ const API_KEY = import.meta.env.VITE_WEATHER_API_KEY;
 
 export default function Weather({ city, state }) {
   const [weather, setWeather] = useState({});
+  
+  const getForecast = async () => {
+    const response = await getWeather(city, state);
 
-  const limit = 1;
-
-
-  const getWeather = async () => {
-    const { data }= await axios.get(
-      `http://api.openweathermap.org/geo/1.0/direct?q=${city},${state}&limit=${limit}&appid=${API_KEY}`
-    );
-
-    const { data: data2} = await axios.get(
-      `https://api.openweathermap.org/data/3.0/onecall?lat=${data[0].lat}&lon=${data[0].lon}&exclude=hourly&appid=${API_KEY}`
-    );
-    
-    setWeather(data2)
-  }
+    setWeather(response);  
+  };
 
 
   useEffect(() => {
-    getWeather();
-  },[])
+    getForecast();
+  }, [])
 
 
   return (
