--- src/utils/common.js
@@ -1,9 +1,0 @@
-
-
-const getRole = () => {
-  return window.sessionStorage.getItem('role').toLowerCase();
-}
-
-
-
-export default getRole
